Revert "add ro.build.fingerprint to OnePlus"
This reverts commit acc06b584f06ad76229821a9ff7776b1130f87bc.

This commit caused ro.build.fingerprint to be static. This resulted in
that PacketManagerService.java:mIsUpgrade did not detect the first boot
after OTA, therefore the package manager caches were not reset and the
versionName/versionCode of system apps updated in the OTA were not
updated.

Now ro.build.fingerprint is autogenerated with build-specific metadata
in Build.java:deriveFingeprint()